+++ payroll_account/models/__init__.py
@@ -1,3 +1,7 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from . import hr_payroll_account
+from . import hr_contract
+from . import hr_payslip_line
+from . import hr_payslip_run
+from . import hr_payslip
+from . import hr_salary_rule
