+++ payroll_account/models/hr_payslip_run.py
@@ -1,0 +1,22 @@
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
+import logging
+
+from odoo import fields, models
+
+logger = logging.getLogger(__name__)
+
+
+class HrPayslipRun(models.Model):
+    _inherit = "hr.payslip.run"
+
+    journal_id = fields.Many2one(
+        "account.journal",
+        "Salary Journal",
+        states={"draft": [("readonly", False)]},
+        readonly=True,
+        required=True,
+        default=lambda self: self.env["account.journal"].search(
+            [("type", "=", "general")], limit=1
+        ),
+    )
